Add a Markdown Code Parser
This commit introduces an alternative parser to the 'doctest' parser.
This parser directly extracts code from Markdown code blocks.  It does
not support adding in expected results to the original document like
the 'doctest' parser, so it is less useful to people reading the source
file.  However, it can prove useful for running properly written
existing readme files and the like.